refactor: clean up copy.ts (WIP) TDE-1551
